increase scope of some variables, define them once
i changed the variables width, height, shape and device to be a global variable, only defined once, instead of in every function. it only gets a value assigned once, in the onlayout function. this should save battery, and will increase efficienty. also, instead of drawing a giant rectangle to blank everything out, i changed that to the clear() function, this is the "proper" way to do that